ssc_analytics_plugin: Warning fix
<==== Mercurial data ====>
Branch: meta_4.0

+++ vms/server/plugins/analytics/ssc_analytics_plugin/src/nx/vms_server_plugins/analytics/ssc/engine.cpp
@@ -39,7 +39,7 @@ static const QByteArray defaultConfiguration = /*suppress newline*/1 + R"json(
 )json";
 
 // The following constants are from "SSC System Specification" (page 5).
-static const unsigned int kCommandLength = 4;
+static const int kCommandLength = 4;
 static const char kSTX = 0x02; //< start byte
 static const char kETX = 0x03; //< stop byte
 static const char kMinDigitCode = 0x30;
@@ -70,7 +70,7 @@ int extractLogicalId(const QByteArray& data)
 /** Check if data begins with a correct command. */
 bool isCorrectCommand(const QByteArray& data)
 {
-    NX_ASSERT(data.size() >= kCommandLength);
+    NX_ASSERT((int) data.size() >= kCommandLength);
     const bool commandEnvelopeIsCorrect = (data[0] == kSTX && data[3] == kETX);
 
     const char b1 = data[1];
